Skeleton Class placing bones together
